Mac/requirements.txt
# Core Dependencies
flask==3.0.0
pyyaml==6.0.1
numpy==1.26.2
scipy==1.11.4
sounddevice==0.4.6
pyperclip==1.8.2

# LLM Provider Support
litellm==1.17.9

# Speech-to-Text Engines
openai-whisper==20231117
faster-whisper==0.10.0
google-cloud-speech==2.24.0
openai==1.6.1
SpeechRecognition==3.10.1

# Windows-specific
pywin32==306; sys_platform == 'win32'
pyaudio==0.2.14; sys_platform == 'win32'

# Optional but recommended
python-dotenv==1.0.0
